+++ src/main/java/com/yandex/filmorate/model/db/GenreRepository.java
@@ -3,14 +3,14 @@ package com.yandex.filmorate.model.db;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface GenreRepository extends JpaRepository<GenreEntity,Long> {
+public interface GenreRepository extends JpaRepository<GenreEntity,Integer> {
 
     @Query(value = """
      SELECT ge.id 
      FROM GenreEntity ge
      WHERE ge.name = :name
     """)
-    Long findByName(String name);
+    Integer findByName(String name);
 
     boolean existsByName(String name);
 }
